pyproject.toml
[project]
name = "paperless-pre-consume-ocr"
version = "1.0.0"
description = "OCR pre-processing script for Paperless-NGX"
requires-python = ">=3.10"
dependencies = [
    "ocrmypdf>=15.0",
    "Pillow>=10.0",
    "img2pdf>=0.5",
    "pikepdf>=8.0",
    "pdfminer.six>=20221105",
    "psycopg[binary]>=3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.6",
    "mypy>=1.10",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = [
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=xml",
]

[tool.ruff]
line-length = 100
target-version = "py310"
src = ["src"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line length handled by formatter
]

[tool.mypy]
python_version = "3.10"
files = ["src"]
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
no_implicit_optional = true
